requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
certifi==2018.1.18
chardet==3.0.4
cryptography==2.1.4
Django==3.1.3
django-filter==21.1
djangorestframework==3.12.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
mysqlclient==2.0.3
pycrypto==2.6.1
PyGObject==3.26.1
pytz==2021.3
pyxdg==0.25
requests==2.18.4
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.4.2
ssh-import-id==5.7
typing-extensions==3.10.0.2
urllib3==1.22
wrapt==1.13.3
